Refactored for initializer only.
