+++ apps/brreg-stub/src/main/java/no/nav/brregstub/BrregApplicationLauncher.java
@@ -1,13 +1,14 @@
 package no.nav.brregstub;
 
 import no.nav.dolly.libs.nais.NaisEnvironmentApplicationContextInitializer;
-import no.nav.dolly.libs.nais.NaisVaultKeyInitializer;
+import no.nav.dolly.libs.vault.NaisVaultKeyInitializer;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.builder.SpringApplicationBuilder;
 
 @SpringBootApplication
 public class BrregApplicationLauncher {
     public static void main(String[] args) {
+        NaisVaultKeyInitializer.run();
         new SpringApplicationBuilder(BrregApplicationLauncher.class)
                 .initializers(new NaisEnvironmentApplicationContextInitializer())
                 .run(args);
