+++ libs/vault/src/main/java/no/nav/dolly/libs/vault/NaisVaultKeyInitializer.java
@@ -1,32 +1,36 @@
-package no.nav.dolly.libs.nais;
+package no.nav.dolly.libs.vault;
 
 import lombok.extern.slf4j.Slf4j;
 
 import java.nio.file.Files;
 import java.nio.file.Paths;
 
+/**
+ * Set the Vault token system property from a file mounted by NAIS.
+ * <br/>
+ * {@code ApplicationContextInitializer} or {@code EnvironmentPostProcessor} cannot be used,
+ * as they are invoked <em>after</em> the Vault configuration has been loaded.
+ */
 @Slf4j
 public class NaisVaultKeyInitializer {
 
-    private static final String KEY = "SPRING_CLOUD_VAULT_TOKEN";
     private static final String PROPERTY = "spring.cloud.vault.token";
     private static final String PATH = "/var/run/secrets/nais.io/vault/vault_token";
 
-    public static void run() {
+    public static void run()
+        throws IllegalStateException {
 
         try {
             var path = Paths.get(PATH);
             if (Files.exists(path)) {
                 var value = Files.readString(path).trim();
-                System.setProperty(KEY, value);
                 System.setProperty(PROPERTY, value);
-                var maskedValue = value.charAt(0) + "*".repeat(value.length() - 2) + value.charAt(value.length() - 1);
-                log.info("System property {} set from file {} with value {}", KEY, PATH, maskedValue);
+                log.info("Vault token set from file {}", PATH);
             } else {
                 log.warn("File not found at {}; hopefully you're running locally", PATH);
             }
         } catch (Exception e) {
-            throw new RuntimeException("Error setting system property %s from file %s".formatted(KEY, PATH), e);
+            throw new IllegalStateException("Error setting Vault token from file %s".formatted(PATH), e);
         }
 
     }
