--- libs/vault/src/main/resources/META-INF/spring.factories
@@ -1,1 +1,0 @@
-org.springframework.boot.env.EnvironmentPostProcessor=no.nav.dolly.libs.vault.VaultKeyInitializer
